🟡(change): notification listener
move the notification listener from the notification menu into the global layout.
