+++ client/src/components/layout/notification_menu/notification-menu.tsx
@@ -17,8 +17,6 @@ import {
 } from "@/components/ui/tooltip";
 import { useNotifications } from "@/hooks/useQueries";
 import axios from "@/lib/axiosConfig";
-import { WEB_SOCKET_URL } from "@/lib/constants";
-import { createWebsocketManager } from "@/lib/websockets";
 import { useUserStore } from "@/stores/AuthStore";
 import { useHeaderStore } from "@/stores/HeaderStore";
 import { UserNotification } from "@/types/accounts";
@@ -121,7 +119,6 @@ export function NotificationMenu() {
     useState<boolean>(false);
   const { id: userId } = useUserStore.get("user");
   const { notificationsData, notificationsLoading } = useNotifications(userId);
-  const webSocketManager = createWebsocketManager();
   const [activeTab, setActiveTab] = useState<string>("inbox");
 
   const markedAndInvalidate = async () => {
@@ -143,36 +140,6 @@ export function NotificationMenu() {
 
     setNotificationMenuOpen(false);
   };
-
-  React.useEffect(() => {
-    webSocketManager.connect("notifications", `${WEB_SOCKET_URL}/${userId}`, {
-      onOpen: () => console.info("Connected to notifications websocket"),
-      onMessage: (event: MessageEvent) => {
-        const data = JSON.parse(event.data);
-        queryClient
-          .invalidateQueries({
-            queryKey: ["userNotifications", userId],
-          })
-          .then(() => {
-            toast.success(
-              <div className="flex flex-col space-y-1">
-                <span className="font-semibold">New Report Available!</span>
-                <span className="text-xs">{data.content}</span>
-              </div>,
-            );
-          });
-      },
-      onClose: (event: CloseEvent) => console.info(event),
-    });
-
-    if (!userId) {
-      return;
-    }
-
-    return () => {
-      webSocketManager.disconnect("notifications");
-    };
-  }, [queryClient, userId, webSocketManager]);
 
   React.useEffect(() => {
     if (
